+++ Packages/com.yontalane.query/Runtime/IQueryUI.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Yontalane.Query
+{
+    public interface IQueryUI
+    {
+        public void Initialize(string text, string description, string[] responses, int initialSelection, Action<QueryEventData> callback, Action<QueryEventData> selectCallback);
+    }
+}